creating mimir lib
