+++ Cargo.toml
@@ -19,11 +19,15 @@ osmpbfreader = "*"
 serde_macros = { version = "*", optional = true }
 
 [features]
-default = ["serde_codegen", "rs-es/default", "skojig/default"]
-nightly = ["serde_macros", "rs-es/nightly", "skojig/nightly"]
+default = ["serde_codegen", "rs-es/default", "skojig/default", "mimir/default"]
+nightly = ["serde_macros", "rs-es/nightly", "skojig/nightly", "mimir/nightly"]
 
 [dependencies.skojig]
 path = "libs/skojig"
+default-features = false
+
+[dependencies.mimir]
+path = "libs/mimir"
 default-features = false
 
 [dev-dependencies.docker_wrapper]
