+++ libs/mimir/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "mimir"
+version = "0.0.1"
+authors = ["Canal TP <dt.ro@canaltp.fr>"]
+build = "build.rs"
+
+[dependencies]
+log = "*"
+env_logger = "*"
+docopt = "*"
+rustc-serialize = "*"
+curl = "*"
+rs-es = { version = "*", default-features = false }
+serde = "*"
+serde_json = "*"
+regex = "*"
+serde_macros = { version = "*", optional = true }
+
+[features]
+default = ["serde_codegen", "rs-es/default"]
+nightly = ["serde_macros", "rs-es/nightly"]
+
+[build-dependencies]
+serde_codegen = { version = "*", optional = true }
+syntex = "*"
