+++ libs/mimir/src/objects.rs.in
